fastapi/requirements.txt
# FastAPI + Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.3.0
langchain-mcp>=0.1.0

# MCP Client
mcp>=1.0.0

# HTTP Client
httpx>=0.27.0

# Data Validation
pydantic>=2.9.0

# Date/Time
python-dateutil>=2.9.0

# CORS
starlette>=0.41.0

# Dev
ruff>=0.7.0
